requirements.txt
cloudpickle==2.0.0
gym==0.21.0
mlflow==1.23.1
numpy==1.19.5
psutil==5.8.0
PyYAML==6.0
scipy==1.5.4 
torch==1.10.0
torch-vision==0.1.6.dev0
tqdm==4.63.0
websocket-client==1.3.1
